MAINT: update license/manifest/readme, move calendar_helpers.py
