+++ MANIFEST.in
@@ -2,6 +2,6 @@ include LICENSE
 
 include etc/requirements*.txt
 
-include trading_calendars/resources *.csv
+include tests/resources *.csv
 include versioneer.py
 include trading_calendars/_version.py
