+++ setup.cfg
@@ -4,6 +4,7 @@
 
 [metadata]
 description-file = README.md
+license_file = LICENSE
 
 [versioneer]
 VCS = git
